tests-infrastructure/terraform/stress-test/deploy/workloads/generators/python/requirements.txt
opentelemetry-api==1.31.1
